Fix the issue with timestamps
There was an issue where using per frame comparisons lead to muted
regions not corresponding to the correct time points. This is now taken
care of. All the muted regions are exactly in the ranges they should be
in. This currently has a granularity of 1 second though, so regions
should be expressed to the nearest second. Audio muting should be fully
functional after this commit.