Implement video masking functionality
videoscub can now mask subsections of video. These regions are masked
with the included png files. Right now the black640x368.png file is
the default hard-coded file that will be used. This is because the
sample video file used for testing is 640x368. Our own internal files
are 1280x720. A corresponding black1280x720.png is also found in the
data directory. When using videoscrub on the real video files, the
hard-coded section telling it to use the 640x368 file should be changed
to reflect the new masking file. Functionality to infer the size of the
input video and generate/select the correct masking file will come in
future commits.