Implement a version of videoscrub not tied to GUI
videoscrub-gui.py is the standalone version with a GUI, while the new
videoscrub.py script is a command line utility that can be called from
other programs in a pipeline. This is an early implementation, with
future revisions separating the functionality between the two scripts
even further, having the GUI script call the new videoscrub.py rather
than just borrow its code.